feat(worker-compat): adds explicit path as module_root to bindings call
A use-case of grandiose could be to use the incoming video frames in a Worker within an environment
such as electron with webpack. Loading a Worker with webpack creates an eval statement which breaks
bindings' module path resolver. Passing the module_root option with the full path solves this issue.